requirements.txt
flask-bootstrap
pycountry
flask-wtf
wtforms
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
PyMySQL==1.1.0
